requirements.txt
torch
torchvision
numpy
scipy
opencv-python
Pillow
matplotlib
PyYAML
tqdm
tensorboard
einops
nuscenes-devkit
pyquaternion
shapely
addict
packaging
pytest
pytest-cov
